.history/requirements_20240620220048.txt
numpy
gensim
scikit-learn
scipy
librosa
matplotlib
pyro-ppl